Add status and refresh subcommands
The status subcommand will list information about the SSH tunnel and the proxy
setting state. If the SSH tunnel is on, it will include information about the
process from ps. This is meant to be used if you want to know the status of the
things involved in the proxy.

The refresh subcommand tears down the existing SSH tunnel (if it exists) and
starts up a new one. This is useful for if you had the socks proxy running, but
lost connection or closed your laptop or something, and you try to access the
internet. Now, you can just run socks refresh and the tunnel will be replaced
with a working one.