Don't try and find process if pidfile is empty
This was a bug that would occur when running socks off with no pid in the
pidfile. The script would try an pass an empty string to `ps -p`, which would
cause an error.

The script now skips over the "is the tunnel still running?" step if there is
no pidfile to check.